requirements.txt
boto3~=1.34.0
bcrypt~=3.2.2
coverage~=7.4.4
iso8601~=2.1.0
Jinja2~=3.1.3
mock~=5.1.0
nose~=1.3.7
pyzmq~=26.0.3
redis~=5.0.1
sphinx~=7.2.6
tornado~=6.4.1
websocket-client~=1.7.0
Django~=5.2.0
djangorestframework~=3.15.1
drf-nested-routers~=0.93.5
djangorestframework-simplejwt~=5.3.1
PyJWT~=2.8.0
django-cors-headers~=4.4.0
pytz~=2024.2
requests~=2.32.3
psycopg2~=2.9.9
sentry-sdk~=1.40.0
gunicorn~=21.2.0
supervisor~=4.2.5
openai~=1.65.4
croniter~=2.0.5
markupsafe~=2.1.5
jsonschema~=4.21.1
kubernetes~=29.0.0
celery[librabbitmq]~=5.3.6
fastapi~=0.111.0
uvicorn~=0.29.0
websockets~=12.0
memcache~=0.12.0
python-json-logger~=2.0.7
pymemcache~=4.0.0
google-auth~=2.29.0
drf-spectacular==0.28.0
httpx~=0.27.0
pydantic~=2.10.0
